@@ -2,8 +2,7 @@ name: pixorama_client
 description: Starting point for a Serverpod client.
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=3.0.0 <4.0.0'
 
 dependencies:
-  serverpod_client:
-    path: ../vendor/serverpod/packages/serverpod_client
+  serverpod_client: 1.2.3
