@@ -4,15 +4,15 @@
 // ignore_for_file: library_private_types_in_public_api
 // ignore_for_file: public_member_api_docs
 // ignore_for_file: implementation_imports
+// ignore_for_file: use_super_parameters
+// ignore_for_file: type_literal_in_constant_pattern
 
 library protocol; // ignore_for_file: no_leading_underscores_for_library_prefixes
 
 import 'package:serverpod/serverpod.dart' as _i1;
-import 'image_data.dart' as _i2;
-import 'image_update.dart' as _i3;
-import 'package:serverpod/protocol.dart' as _i4;
-export 'image_data.dart';
-export 'image_update.dart'; // ignore_for_file: equal_keys_in_map
+import 'package:serverpod/protocol.dart' as _i2;
+import 'example.dart' as _i3;
+export 'example.dart';
 
 class Protocol extends _i1.SerializationManagerServer {
   Protocol._();
@@ -23,6 +23,10 @@ class Protocol extends _i1.SerializationManagerServer {
 
   static final Protocol _instance = Protocol._();
 
+  static final List<_i2.TableDefinition> targetTableDefinitions = [
+    ..._i2.Protocol.targetTableDefinitions
+  ];
+
   @override
   T deserialize<T>(
     dynamic data, [
@@ -32,42 +36,30 @@ class Protocol extends _i1.SerializationManagerServer {
     if (customConstructors.containsKey(t)) {
       return customConstructors[t]!(data, this) as T;
     }
-    if (t == _i2.ImageData) {
-      return _i2.ImageData.fromJson(data, this) as T;
+    if (t == _i3.Example) {
+      return _i3.Example.fromJson(data, this) as T;
     }
-    if (t == _i3.ImageUpdate) {
-      return _i3.ImageUpdate.fromJson(data, this) as T;
-    }
-    if (t == _i1.getType<_i2.ImageData?>()) {
-      return (data != null ? _i2.ImageData.fromJson(data, this) : null) as T;
-    }
-    if (t == _i1.getType<_i3.ImageUpdate?>()) {
-      return (data != null ? _i3.ImageUpdate.fromJson(data, this) : null) as T;
+    if (t == _i1.getType<_i3.Example?>()) {
+      return (data != null ? _i3.Example.fromJson(data, this) : null) as T;
     }
     try {
-      return _i4.Protocol().deserialize<T>(data, t);
+      return _i2.Protocol().deserialize<T>(data, t);
     } catch (_) {}
     return super.deserialize<T>(data, t);
   }
 
   @override
   String? getClassNameForObject(Object data) {
-    if (data is _i2.ImageData) {
-      return 'ImageData';
-    }
-    if (data is _i3.ImageUpdate) {
-      return 'ImageUpdate';
+    if (data is _i3.Example) {
+      return 'Example';
     }
     return super.getClassNameForObject(data);
   }
 
   @override
   dynamic deserializeByClassName(Map<String, dynamic> data) {
-    if (data['className'] == 'ImageData') {
-      return deserialize<_i2.ImageData>(data['data']);
-    }
-    if (data['className'] == 'ImageUpdate') {
-      return deserialize<_i3.ImageUpdate>(data['data']);
+    if (data['className'] == 'Example') {
+      return deserialize<_i3.Example>(data['data']);
     }
     return super.deserializeByClassName(data);
   }
@@ -75,15 +67,18 @@ class Protocol extends _i1.SerializationManagerServer {
   @override
   _i1.Table? getTableForType(Type t) {
     {
-      var table = _i4.Protocol().getTableForType(t);
+      var table = _i2.Protocol().getTableForType(t);
       if (table != null) {
         return table;
       }
     }
-    switch (t) {
-      case _i2.ImageData:
-        return _i2.ImageData.t;
-    }
     return null;
   }
+
+  @override
+  List<_i2.TableDefinition> getTargetTableDefinitions() =>
+      targetTableDefinitions;
+
+  @override
+  String getModuleName() => 'pixorama';
 }
